update: extraテストを追加 & Doにも失敗時のリトライログを追加
